[refactoring] Refactored out logging and options to simplify crawlerFactory.js
