+++ lib/crawlerFactory.js
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-const mockInsights = require('../providers/logger/mockInsights');
 const LimitedTokenFactory = require('../providers/fetcher/limitedTokenFactory');
 const TokenFactory = require('../providers/fetcher/tokenFactory');
 const ComputeLimiter = require('../providers/limiting/computeLimiter');
@@ -19,18 +18,14 @@ const AzureStorageDocStore = require('../providers/storage/storageDocStore');
 const UrlToUrnMappingStore = require('../providers/storage/urlToUrnMappingStore');
 const AzureTableMappingStore = require('../providers/storage/tableMappingStore');
 const amqp10 = require('amqp10');
-const appInsights = require('applicationinsights');
-const aiLogger = require('winston-azure-application-insights').AzureApplicationInsightsLogger;
 const AzureStorage = require('azure-storage');
 const config = require('painless-config');
 const Crawler = require('../lib/crawler');
 const CrawlerService = require('../lib/crawlerService');
-const fs = require('fs');
 const GitHubFetcher = require('../providers/fetcher/githubFetcher');
 const GitHubProcessor = require('../providers/fetcher/githubProcessor');
 const ip = require('ip');
 const moment = require('moment');
-const policy = require('../lib/traversalPolicy');
 const Q = require('q');
 const QueueSet = require('../providers/queuing/queueSet');
 const redis = require('redis');
@@ -42,80 +37,16 @@ const RefreshingConfigRedis = require('refreshing-config-redis');
 const request = require('request');
 const Request = require('../lib/request');
 const requestor = require('ghrequestor');
-const winston = require('winston');
+const createLogger = require('./logging').createLogger;
+const getDefaultOptions = require('./options').getDefaultOptions;
 
 const AmqpClient = amqp10.Client;
 const AmqpPolicy = amqp10.Policy;
 
-let factoryLogger = null;
+let factoryLogger = createLogger(true);
 let redisClient = null;
 
 class CrawlerFactory {
-
-  static getDefaultOptions() {
-    return {
-      crawler: {
-        name: config.get('CRAWLER_NAME') || 'crawler',
-        count: 0,
-        pollingDelay: 5000,
-        processingTtl: 60 * 1000,
-        promiseTrace: false,
-        requeueDelay: 5000,
-        orgList: CrawlerFactory.loadOrgs(),
-        deadletterPolicy: 'always' // Another option: excludeNotFound
-      },
-      fetcher: {
-        tokenLowerBound: 50,
-        metricsStore: 'redis',
-        callCapStore: 'memory',
-        callCapWindow: 1,       // seconds
-        callCapLimit: 30,       // calls
-        computeLimitStore: 'memory',
-        computeWindow: 15,      // seconds
-        computeLimit: 15000,    // milliseconds
-        baselineFrequency: 60,  // seconds
-        deferDelay: 500
-      },
-      queuing: {
-        provider: config.get('CRAWLER_QUEUE_PROVIDER') || 'amqp10',
-        queueName: config.get('CRAWLER_QUEUE_PREFIX') || 'crawler',
-        credit: 100,
-        weights: { events: 10, immediate: 3, soon: 2, normal: 3, later: 2 },
-        messageSize: 240,
-        parallelPush: 10,
-        pushRateLimit: 200,
-        metricsStore: 'redis',
-        events: {
-          provider: config.get('CRAWLER_EVENT_PROVIDER') || 'webhook',
-          topic: config.get('CRAWLER_EVENT_TOPIC_NAME') || 'crawler',
-          queueName: config.get('CRAWLER_EVENT_QUEUE_NAME') || 'crawler'
-        },
-        attenuation: {
-          ttl: 3000
-        },
-        tracker: {
-          // driftFactor: 0.01,
-          // retryCount: 3,
-          // retryDelay: 200,
-          // locking: true,
-          // lockTtl: 1000,
-          ttl: 60 * 60 * 1000
-        }
-      },
-      storage: {
-        ttl: 3 * 1000,
-        provider: config.get('CRAWLER_STORE_PROVIDER') || 'azure',
-        delta: {
-          provider: config.get('CRAWLER_DELTA_PROVIDER')
-        }
-      },
-      locker: {
-        provider: 'redis',
-        retryCount: 3,
-        retryDelay: 200
-      }
-    };
-  }
 
   static createService(name) {
     factoryLogger.info('appInitStart');
@@ -158,7 +89,7 @@ class CrawlerFactory {
 
   static _decorateOptions(options) {
     Object.getOwnPropertyNames(options).forEach(key => {
-      const logger = CrawlerFactory.createLogger(true);
+      const logger = createLogger(true);
       options[key].logger = logger;
       const capitalized = key.charAt(0).toUpperCase() + key.slice(1);
       const metricsFactory = CrawlerFactory[`create${capitalized}Metrics`];
@@ -204,7 +135,7 @@ class CrawlerFactory {
       }
       return config.getAll().then(values => {
         factoryLogger.info(`creating refreshingOption config get completed`);
-        const defaults = CrawlerFactory.getDefaultOptions();
+        const defaults = getDefaultOptions();
         return CrawlerFactory.initializeSubsystemOptions(values, defaults[subsystemName]).then(resolved => {
           factoryLogger.info(`subsystem options initialized`);
           result[subsystemName] = values;
@@ -224,7 +155,7 @@ class CrawlerFactory {
 
   static createRedisRefreshingConfig(crawlerName, subsystemName) {
     factoryLogger.info('Create refreshing redis config', { crawlerName: crawlerName, subsystemName: subsystemName });
-    const redisClient = CrawlerFactory.getRedisClient(CrawlerFactory.createLogger(true));
+    const redisClient = CrawlerFactory.getRedisClient(createLogger(true));
     const key = `${crawlerName}:options:${subsystemName}`;
     const channel = `${key}-channel`;
     const configStore = new RefreshingConfigRedis.RedisConfigStore(redisClient, key);
@@ -283,7 +214,7 @@ class CrawlerFactory {
   static createTokenLimiter(options) {
     factoryLogger.info('create token limiter', { capStore: options.capStore });
     return options.capStore === 'redis'
-      ? CrawlerFactory.createRedisTokenLimiter(getRedisClient(options.logger), options)
+      ? CrawlerFactory.createRedisTokenLimiter(CrawlerFactory.getRedisClient(options.logger), options)
       : CrawlerFactory.createInMemoryTokenLimiter(options);
   }
 
@@ -527,18 +458,6 @@ class CrawlerFactory {
     });
   }
 
-  static createLogger(echo = false, level = 'info') {
-    mockInsights.setup(config.get('CRAWLER_INSIGHTS_KEY') || 'mock', echo);
-    const result = new winston.Logger();
-    result.add(aiLogger, {
-      insights: appInsights,
-      treatErrorsAsExceptions: true,
-      exitOnError: false,
-      level: level
-    });
-    return result;
-  }
-
   static createRequestTracker(prefix, options) {
     let locker = null;
     if (options.tracker.locking) {
@@ -652,28 +571,7 @@ class CrawlerFactory {
     });
     return queuesMetrics;
   }
-
-  static loadOrgs() {
-    let orgList = config.get('CRAWLER_ORGS');
-    if (orgList) {
-      orgList = orgList.split(';').map(entry => entry.toLowerCase().trim());
-    } else {
-      orgList = CrawlerFactory._loadLines(config.get('CRAWLER_ORGS_FILE'));
-    }
-    return orgList;
-  }
-
-  static _loadLines(path) {
-    if (!path || !fs.existsSync(path)) {
-      return [];
-    }
-    let result = fs.readFileSync(path, 'utf8');
-    result = result.split(/\s/);
-    return result.filter(line => { return line; }).map(line => { return line.toLowerCase(); });
-  }
 }
-
-factoryLogger = CrawlerFactory.createLogger(true);
 
 module.exports = CrawlerFactory;
 
@@ -686,7 +584,6 @@ class EventFormatter {
   format(message) {
     // The message here is expected to be a WEBHOOK event.  Use the information included to identify the
     // repo or org to poll for new events.
-    const type = message.applicationProperties.event;
     const event = message.body;
     const eventsUrl = event.repository ? event.repository.events_url : event.organization.events_url;
     const result = new Request('event_trigger', `${eventsUrl}`);
